Create monitor.bas
this is a basic machine code monitor for c64. written using cbm prg studio. It has simple menu consisting of 1) peek 2) poke 3) scroll
the peek and poke ask for number of bytes to work on, poke also asks for a new value, scroll prints a full screen from a start address in binary, and "e" allows the current memeory pointer (indicated by <=) to be edited. All the displays print first the memory address in decimal (hope to work out how to print it in hex! and then the memory value in binary, where lsb=0 so 8 decimal = 0001000